+++ .github/workflows/test.yml
@@ -26,7 +26,7 @@ jobs:
       - name: Select Xcode
         uses: maxim-lobanov/setup-xcode@v1
         with:
-          xcode-version: "16.3.0"
+          xcode-version: latest-stable
       - name: Cache SwiftPM
         uses: actions/cache@v4
         with:
