+++ Sources/HealthChecksMocks/Mocks/Constants.swift
@@ -16,17 +16,20 @@
 //  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 //
-//  MongoRequestMock.swift
+//  Constants.swift
 //
 //
 //  Created by Mykola Buhaiov on 15.03.2024.
 //
 
 import Vapor
-@testable import HealthChecks
 
-public struct MongoRequestMock: MongoRequestSendable {
-    public func getConnection(by url: String) async -> String {
-        "connecting"
-    }
+public enum Constants {
+    // swiftlint:disable numbers_smell
+    public static let psqlVersionDescription =
+        """
+        PostgreSQL 14.10 on x86_64-pc-linux-musl, compiled by gcc (Alpine 13.2.1_git20231014) 13.2.1 20231014, 64-bit
+        """
+    public static let consulUrl = "http://127.0.0.1:8500"
+    // swiftlint:enable numbers_smell
 }
