+++ Sources/HealthChecksMocks/Mocks/MongoRequestMock.swift
@@ -16,21 +16,24 @@
 //  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 //
-//  Constants.swift
+//  MongoRequestMock.swift
 //
 //
 //  Created by Mykola Buhaiov on 15.03.2024.
 //
 
 import Vapor
+@testable import HealthChecks
 
+public struct MongoRequestMock: MongoRequestSendable {
+    // Example of value: `connecting`, `disconnected`, `connected(connectionCount: Int)`, `closed`
+    private var connection: String
+    
+    public init(connection: String = "connecting") {
+        self.connection = connection
+    }
 
-enum Constants {
-    // swiftlint:disable numbers_smell
-    static let psqlVersionDescription =
-        """
-        PostgreSQL 14.10 on x86_64-pc-linux-musl, compiled by gcc (Alpine 13.2.1_git20231014) 13.2.1 20231014, 64-bit
-        """
-    static let consulUrl = "http://127.0.0.1:8500"
-    // swiftlint:enable numbers_smell
+    public func getConnection(by url: String) async -> String {
+        self.connection
+    }
 }
