+++ Sources/HealthChecksMocks/Mocks/RedisRequestMock.swift
